+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1 - Primeiro release
+## 0.0.1
 
-* Primeiro release
+* First release
+* Instagram Share to Story for Android
+* Facebook Share to Story for Android
+* Facebook Share to Reels for Android
+
+## 0.0.2
+
+* Update README
+* Update EXAMPLE